requirements.txt
Django==4.2.7
asgiref==3.7.2
pandas==1.5.3
numpy==1.24.3
python-decouple==3.8
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
Pillow==10.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
daphne==4.0.0
